+++ doc/RELEASE-NOTES.md
@@ -2,6 +2,8 @@
 
 ## 0.0.26
 
+* Add 'retz.exectutor.java' to server configuration file for more flexible
+  Java installation on Mesos Agents
 * Add 'retz.max.running' to server configuration file. This number limits
   number of simultaneously running jobs, to control Retz load to whole
   Mesos cluster.